Scene Save, Load 구현중( 컴포넌트들) 예제별 TestScene 만드는중.
string 데이터가 깨져서 읽혀지는 문제

+++ Project/Engine/Function.cpp
@@ -3,16 +3,16 @@
 
 void SaveStringToFile(const tstring& _str, FILE* _pFile)
 {
-	int iLen = _str.length();
-	fwrite(&iLen, sizeof(int), iLen, _pFile);
+	UINT iLen = (UINT)_str.length();
+	fwrite(&iLen, sizeof(UINT), iLen, _pFile);
 	fwrite(_str.data(), sizeof(TCHAR), iLen, _pFile);
 }
 
 void LoadStringFromFile(tstring& _str, FILE* _pFile)
 {
 	TCHAR szBuffer[1024] = {};
-	int iLen = 0;
-	fread(&iLen, sizeof(int), 1, _pFile);
+	UINT iLen = 0;
+	fread(&iLen, sizeof(UINT), 1, _pFile);
 	fread(szBuffer, sizeof(TCHAR), iLen, _pFile);
 	_str = szBuffer;
 }
